Revert _rlnOpticsGroup type after altering example starfile.

+++ src/aspire/utils/relion_interop.py
@@ -44,7 +44,7 @@ relion_metadata_fields = {
     "_rlnNrOfSignificantSamples": float,
     "_rlnNrOfFrames": int,
     "_rlnMaxValueProbDistribution": float,
-    "_rlnOpticsGroup": float,
+    "_rlnOpticsGroup": int,
     "_rlnOpticsGroupName": str,
 }
